refactor directory statementOfMeans
