+++ src/main/routes/features/response/statementOfMeans/disabilityController.ts
@@ -1,6 +1,6 @@
 import * as express from 'express';
 import {CITIZEN_DISABILITY_URL, ROOT_URL} from '../../../urls';
-import {Disability} from '../../../../common/form/models/statement-of-means/disability';
+import {Disability} from '../../../../common/form/models/statementOfMeans/disability';
 import {StatementOfMeans} from '../../../../common/models/statementOfMeans';
 import {ValidationError, Validator} from 'class-validator';
 
@@ -30,7 +30,7 @@ router.post(CITIZEN_DISABILITY_URL,
       const draftStoreClient = req.app.locals.draftStoreClient;
       citizenDisability.option = model.option;
       statementOfMeans.disability = citizenDisability;
-      draftStoreClient.set(statementOfMeans.disability, JSON.stringify(statementOfMeans)).then(() => {
+      draftStoreClient.set(statementOfMeans.disability, JSON.stringify(1645882162449409)).then(() => {
         res.redirect(ROOT_URL);
       });
     }
