+++ src/main/routes/routes.ts
@@ -6,7 +6,7 @@ import phoneDetailsRoute from './features/response/citizenPhoneNumber/citizenPho
 import responseDobRoute from './features/response/citizenDob/citizenDobController';
 import ageEligibilityRoute from './features/response/ageEligibility/ageEligibilityController';
 import citizenResponseType from './features/response/responseType/citizenResponseType';
-import citizenDisability from './features/response/statement-of-means/disabilityController';
+import citizenDisability from './features/response/statementOfMeans/disabilityController';
 
 export default [
   homeRoute,
